displayed deleted track list
Implemented a way to display the deleted track list on the left side of the screen. I also populated a confirm button is there was at least one track in the deleted list. i also added a button to remove a track from this lsit if the user wants to keep the song. MUST IMPLEMENT A FUNCTION TO DO THIS. currently the x button does nothing.